requirements.txt
python-dotenv==1.0.1
mlxtend==0.23.1
google-cloud-storage==2.18.2
pinecone==5.3.1
pandas==2.0.3
google-generativeai
pydantic-settings==2.4.0  
fastapi==0.110.2
uvicorn[standard]==0.29.0
